Menu driven program of food items
Here, I have displayed a list of food items. The user must enter a value corresponding to the food item, and the program outputs the food item selected along with its price. Here i have used a switch case as the switch case statement is used when we have multiple options and we need to perform a different task for each option.